feat(llm-mindspore): 添加 Transformer 机器翻译和 Bert 文本分类实战代码
- 新增 Transformer 机器翻译实战代码，实现中英文翻译功能
- 新增 Bert 文本分类实战代码，用于商品评论情感分析
- 删除了原有的 Transformer Encoder 实战代码

+++ cpp/LQB/Day_one/PrefixSum.cpp
@@ -22,7 +22,7 @@ int main()
         cin >> a[i];
         for (int j = 1; j <= 5; j++)
         {
-            b[i][j] = b[i - 1][j] + pow(a[i], j);
+            b[i][j] = b[i - 1][j] + static_cast<long long>(pow(a[i], j));
         }
     }
